2003-02-16  Daniel Morgan <danmorg@sc.rr.com>
	* System.Data.OracleClient.Oci/OciStatementHandle.cs
	* System.Data.OracleClient.Oci/OciColumnInfo.cs
	* System.Data.OracleClient/OciGlue.cs
	* System.Data.OracleClient/OracleDataReader.cs: fixes
	  to display character data in SQL# command-line
	  and GTK# versions

	* Test/TestOracleClient.cs: updated test to
	  include reading data via the OracleDataReader

	* System.Data.OracleClient/OracleConnection.cs: removed
	  debug code

svn path=/trunk/mcs/; revision=11614

+++ mcs/class/System.Data.OracleClient/System.Data.OracleClient.Oci/OciColumnInfo.cs
@@ -22,9 +22,9 @@ namespace System.Data.OracleClient.Oci {
 	{
 		public string ColumnName;
 		public int ColumnOrdinal;
-		public int ColumnSize;
+		public ushort ColumnSize;
 		public byte Precision;
-		public byte Scale;
+		public sbyte Scale;
 		public OciDataType DataType;
 		public bool AllowDBNull;
 		public string BaseColumnName;
